refactor(test): move stray property tests into *.property.test.ts files
The nightly property job selects by filename (bun test property.test), so
property assertions left in mixed *.test.ts files were excluded. Relocate them
into sibling property files so every propertyConfig-backed test is covered:

- folio markdown: extract the loose-list idempotence property into
  fromMarkdown.property.test.ts.
- folio paged-layout: move the three paragraph-measurement properties into the
  existing incrementalMeasure.property.test.ts.
- template-conditions: extract the evaluateCondition reference-evaluator
  property into index.property.test.ts.

The example (non-property) tests stay in place.
