+++ packages/folio/src/core/markdown/fromMarkdown.test.ts
@@ -10,9 +10,6 @@
 //    emits (headings, prose, marks, lists, tables, inline code, blockquotes).
 
 import { describe, expect, test } from "bun:test";
-import fc from "fast-check";
-
-import { propertyConfig } from "@stll/property-testing";
 
 import { fromMarkdown } from "./fromMarkdown";
 import { toMarkdown } from "./index";
@@ -147,16 +144,6 @@ describe("markdown bridge — round-trip", () => {
     ).toBe("[site](https://example.com/a) and [section](#intro)");
   });
 
-  test("loose list items are idempotent after markdown bridge normalization", () => {
-    fc.assert(
-      fc.property(
-        looseListItemMarkdown(),
-        (source) => normalize(normalize(source)) === normalize(source),
-      ),
-      propertyConfig({ numRuns: 100 }),
-    );
-  });
-
   test("a soft line break becomes a break node, not a raw newline in a run", () => {
     // A Word run can't carry "\n"; the layout engine renders such lines on top
     // of each other. A two-line blockquote must produce an explicit break.
@@ -185,63 +172,3 @@ describe("markdown bridge — round-trip", () => {
     expect(breakNodes).toBeGreaterThan(0);
   });
 });
-
-const LOWERCASE_WORD_CHARS = [
-  "a",
-  "b",
-  "c",
-  "d",
-  "e",
-  "f",
-  "g",
-  "h",
-  "i",
-  "j",
-  "k",
-  "l",
-  "m",
-  "n",
-  "o",
-  "p",
-  "q",
-  "r",
-  "s",
-  "t",
-  "u",
-  "v",
-  "w",
-  "x",
-  "y",
-  "z",
-] as const;
-
-const inlineText = fc
-  .array(
-    fc.array(fc.constantFrom(...LOWERCASE_WORD_CHARS), {
-      minLength: 1,
-      maxLength: 8,
-    }),
-    { minLength: 1, maxLength: 3 },
-  )
-  .map((words) => words.map((chars) => chars.join("")).join(" "));
-
-const inlineMarkdown = fc.oneof(
-  inlineText,
-  inlineText.map((value) => `**${value}**`),
-  inlineText.map((value) => `*${value}*`),
-  inlineText.map((value) => `\`${value.replaceAll("`", "")}\``),
-  inlineText.map(
-    (value) => `[${value}](https://example.com/${encodeURIComponent(value)})`,
-  ),
-);
-
-const looseListItemMarkdown = () =>
-  fc
-    .tuple(
-      fc.array(inlineMarkdown, { minLength: 1, maxLength: 3 }),
-      fc.array(inlineMarkdown, { minLength: 1, maxLength: 3 }),
-    )
-    .map(
-      ([firstParagraph, secondParagraph]) =>
-        `- ${firstParagraph.join(" ")}\n\n  ${secondParagraph.join(" ")}`,
-    );
